Re-organised the common memory resource tests a little.
Made the tests part of the vecmem_testing_common static library,
moving most of their code into .cpp files.

Removed the references to CUDA from the documentation of the
common tests.
