+++ tests/CMakeLists.txt
@@ -23,7 +23,14 @@ endif()
 
 # Build a common, helper library.
 add_library( vecmem_testing_common STATIC
-   "common/memory_resource_name_gen.hpp" "common/memory_resource_name_gen.cpp" )
+   "common/memory_resource_name_gen.hpp"
+   "common/memory_resource_name_gen.cpp"
+   "common/memory_resource_test_basic.hpp"
+   "common/memory_resource_test_basic.cpp"
+   "common/memory_resource_test_host_accessible.hpp"
+   "common/memory_resource_test_host_accessible.cpp"
+   "common/memory_resource_test_stress.hpp"
+   "common/memory_resource_test_stress.cpp" )
 target_link_libraries( vecmem_testing_common
    PUBLIC vecmem::core GTest::gtest )
